requirements.txt
# Core Framework
quart==0.20.0
hypercorn==0.16.0

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
aiomysql==0.2.0
alembic==1.13.1

# Security
quart-auth==0.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-multipart==0.0.6

# Caching & Sessions
redis==5.0.1
quart-session==3.0.0

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.11.5
pydantic-settings==2.9.1

# Logging & Monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# HTTP & Requests
httpx==0.25.2
aiofiles==23.2.1

# Validation & Serialization
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3